+++ src/ser.rs
@@ -62,6 +62,22 @@ where
     encoder.encode_stream(wrt, &mut msg)
 }
 
+/// Serialize user data into pre-allocated buffer.
+/// Returns number of bytes written.
+#[allow(unused)]
+pub fn to_buffer<T>(encoder: &mut Encoder, buffer: &mut [u8], value: &T) -> Result<usize>
+where
+    T: ?Sized + Serialize,
+{
+    // Serialise user data into internal data model
+    let mut data = TemplateData::new_empty();
+    value.serialize(&mut data)?;
+
+    // Encode FAST message from internal data model
+    let mut msg = ModelVisitor::new(data);
+    encoder.encode_buffer(buffer, &mut msg)
+}
+
 impl serde::ser::Error for Error {
     fn custom<T>(msg: T) -> Self
     where